+++ packages/target-tauri/src-tauri/src/deeplink.rs
@@ -62,6 +62,16 @@ pub async fn handle_deep_link(
             return Ok(());
         }
 
+        #[cfg(target_os = "windows")]
+        if potential_scheme == "dcnotification" {
+            use user_notify::windows::decode_deeplink;
+
+            let response = decode_deeplink(&deeplink_or_xdc)?;
+            crate::Notifications::handle_response(app, response).await;
+
+            return Ok(());
+        }
+
         if potential_scheme != "file" {
             log::error!("handle_deep_link: scheme {potential_scheme}: is unknown")
         }